SPM-BackEnd/requirements.txt
Pillow
opencv-python
numpy
pyvista
tqdm
